requirements.txt
Flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
matplotlib==3.7.1
seaborn==0.12.2
PyPDF2==3.0.1
python-docx==0.8.11
wordcloud==1.9.2
joblib==1.3.2
Werkzeug==2.3.7
